@feat(rf-storage): 存储槽组表支持 RF 槽位 100-131，接收器支持 ADC 通道切换，CLI 槽号 3 位

+++ UserApp/cli.hpp
@@ -34,11 +34,11 @@ public:
 
 private:
     void dispatch();
-    void onLearn(uint8_t slot);
-    void onSend(uint8_t slot);
+    void onLearn(uint16_t slot);
+    void onSend(uint16_t slot);
     void onHelp();
     void onSlots();
-    void onDump(uint8_t slot);   // 诊断：打印槽内全部段(带符号时长μs)
+    void onDump(uint16_t slot);  // 诊断：打印槽内全部段(带符号时长μs)
     void onDbg();                // 诊断：1s 内 ADC 采样 min/max/avg/边沿数
     void onRaw();                // 诊断：固定 200ms 窗口抓边沿(绕过空闲判定)并打印
     void printRaw(const Signal& sig, uint32_t len); // 带符号逗号分隔打印，末尾 len=段数
